Add CC0 license, kopimi text
As Kopimi isn't a recognised license, it makes sense to add a public
domain license like CC0. There are other, similar licenses that
could be used e.g. the [Unlicense][] or [others][].

[Unlicense]: https://choosealicense.com/licenses/#unlicense
[others]: https://choosealicense.com/appendix/

That said, the original author clearly wanted the Kopimi symbol and
link present, so I've added clarifying text and an icon.

I've not deleted the License section as it'll conflict with my other
PR and it's not worth doing if this change isn't correct.